### 2018-10-08 #1004 @rrrlw
@chasemc, @nagoue, @tomsing1, and @giraola appear to have a biochemistry/biology background as well as experience with R; would any be willing to review?
