### 2018-10-08 441f469
Updated JOSS review checklist
