### 2018-10-09 #1004 @nagoue
Hi, I will be happy to review it too.

Le lun. 8 oct. 2018 à 03:37, Raoul W. <notifications@github.com> a écrit :

> @chasemc <https://github.com/chasemc>, @nagoue <https://github.com/nagoue>,
> @tomsing1 <https://github.com/tomsing1>, and @giraola
> <https://github.com/giraola> appear to have a biochemistry/biology
> background as well as experience with R; would any be willing to review?
>
