### 2018-10-09 #1008 @whedon
[REVIEW]: Visualizing alpha-helical peptides in R with helixvis
**Submitting author:** @rrrlw (<a href="http://orcid.org/0000-0003-0503-9580">Raoul R. Wadhwa</a>)
**Repository:** <a href="https://github.com/rrrlw/helixvis" target ="_blank">https://github.com/rrrlw/helixvis</a>
**Version:** v1.0.0
