+++ JOSS-paper/reviewer-checklist.md
@@ -19,4 +19,4 @@
 	- Raoul R Wadhwa: *contributed to code for R package and paper writing, submitting author*
 	- Vigneshwar Subramanian: *contributed to code for Python package and software design*
 	- Regina Stevens-Truss: *contributed to design of features implemented in R package and paper writing, senior author*
-* Novel software: *no known R package implements visualization provided in helixvis*
+* Novel software: *no known R package implements visualization provided in helixvis & this is only programmatic interface authors could find*
