### 2018-10-09 #1008 @rrrlw
First off, thank you @yochannah for editing and @tomsing1, @nagoue for reviewing! Very grateful for the time and effort you all are contributing.

A couple of quick notes for the reviewers:

- The CRAN version of helixvis is very much out-of-date (will update on CRAN when the review on this thread is complete, even if helixvis is not accepted to JOSS), so please use `devtools::install_github("rrrlw/helixvis")` to install the package
- I have added a document reflecting the reviewer checklist for this JOSS submission [here](https://github.com/rrrlw/helixvis/blob/master/JOSS-paper/reviewer-checklist.md), which could save the reviewers some time on the checklists above. Hope it helps.

Looking forward to everyone's feedback!
