### 2018-10-16 22bcd9f
Updated examples in vignette
