### 2018-10-16 a665f3b
Added RST edits to intro vignette
