### 2018-10-21 #1008 @nagoue
Hi,
I have a minor comments:

1. Consider adding **units (>= 0.6-1)** in the Imports part of the DESCRIPTION file as the installation wasn't going smoothly with the devtools option.

2. There is a typo error l40 in paper.md (graphical).

3. Redirection is not working for reference Sandve et al.

as well as improvement suggestions:

4. Include a legend to graphic plots.

5. There is no reason to limit peptide length to 18 aa in wenxiang diagram unless to make it comparable to the helical diagram. I would rather suggest not limiting this length to wenxiang diagram style and "only" give a warning to the user to inform him that the comparison between the 2 diagram styles will not be possible.
