### 2018-10-21 rrrlw/helixvis#7 @tomsing1
Include the sequence.R file in the data directory
The authors include a small set of example peptides sequences, which is great. The source
code includes the data in two forms: the binary `data/sequence.rda` and the plain
`data-raw/sequence.R` file with the code that generates the former.

The `sequence.R` file can be included in the `data` directory directly and will be
sourced automatically when then `data("sequence")` call is executed. There is no need
for the binary file (or the `data-raw` directory). See the `?data` help page for details.

When the `sequence.R` file is placed into the `data` directory, it is sourced and there is no need for the `devtools::use_data(sequence)` call.

https://github.com/rrrlw/helixvis/blob/33103cd9d6cbff7cc3374fb92f312838f564e747/data-raw/sequence.R#L17
