### 2018-10-21 rrrlw/helixvis#8 @tomsing1
Mismatch between version in the DESCRIPTION and github release
The DESCRIPTION file currently lists version `1.0.0` but the last github release is `0.9`.

https://github.com/rrrlw/helixvis/blob/33103cd9d6cbff7cc3374fb92f312838f564e747/DESCRIPTION#L4
