### 2018-10-21 rrrlw/helixvis#9 @tomsing1
Include 
To make sure users who install the package from github see the great vignette, it would be useful to include the `build_vignettes=TRUE` option in the `install_github` code provided in the `README.md` file, e.g.

```r
devtools::install_github("rrrlw/helixvis", build_vignettes=TRUE)
```
