### 2018-10-21 #1008 @tomsing1
Dear authors and editors,

I have completed my review. Please find a few minor suggestions below (also filed as issues in the `helixvis` repository). Great work!

## Installation: Does installation proceed as outlined in the documentation?

[Issue #9](https://github.com/rrrlw/helixvis/issues/9)
Yes. The authors included a great vignette with the package - but it is not built automatically when the package is pulled from github. I recommend including

```r
devtools::install_github("rrrlw/helixvis", build_vignettes=TRUE)
````

in the `README.md` file to make sure all users get to see the compiled vignette.

## Does the release version given match the GitHub release (v1.0.0)?

[Issue #8](https://github.com/rrrlw/helixvis/issues/8)
No, the current github release is v0.9, but the DESCRIPTION file shows v1.0.0.
There have been > 90 commits to the master branch since v0.9 has been released.
(I have reviewed the master branch by cloning the master branch.)

## Authorship: Does the full list of paper authors seem appropriate and complete?

rrrlw has currently made all 106 commits to the github repository, so I cannot verify the contributions of the other authors outlined in [additional information provided by the authors](https://github.com/rrrlw/helixvis/blob/master/JOSS-paper/reviewer-checklist.md). I will therefore trust that

> The authors themselves assume responsibility for deciding who should be credited with co-authorship.

as outlined in the [reviewer guidelines](https://joss.readthedocs.io/en/latest/submitting.html#authorship).

## Example usage: Do the authors include examples of how to use the software (ideally to solve real-world analysis problems).

[Issue #7](https://github.com/rrrlw/helixvis/issues/7)
The authors include a small set of example peptides sequences, which is great. The source code includes the data in two forms: the binary `data/sequence.rda` and the plain `data-raw/sequence.R` file with the code that generates the former.

The `sequence.R` file can be included in the `data` directory directly and will be sourced automatically when then `data("sequence")` call is executed. There is no need for the binary file (or the `data-raw` directory). See the `?data` help page for details.

## Code style

The code is very clear, but the authors might want to consider automatic checking of the
code style, e.g.

```r
library(lintr)
lintr::lint_package()
```
and automatic styling e.g. using

```r
library(styler)
styler::style_pkg()
```
