### 2018-10-21 557bcd0
Users following README instructions for GitHub download will get vignette
