### 2018-10-21 rrrlw/helixvis#9 @rrrlw
Thank you for the suggestion! The change has been made (commit 557bcd0).
