### 2018-10-22 d7e6cdb
Removes need for data-raw directory; fixes issue #7
