### 2018-10-22 rrrlw/helixvis#7 @rrrlw
I didn't realize that `.R` files were automatically sourced. The redundant files/directories have been removed and `sequence.R` has been placed in the `data` directory (commit d7e6cdb). Thank you!
