### 2018-10-23 d99821a
Fixed typo in JOSS paper (JOSS review)
