### 2018-10-23 33aa3df
Updated Travis to check code style with lintr
