### 2018-10-23 e8fe3d3
Updated Travis YML to turn off lintr comment bot
