+++ .travis.yml
@@ -12,6 +12,9 @@ r_github_packages:
   - r-lib/covr
   - jimhester/lintr
 
+env:
+  - LINTR_COMMENT_BOT=false
+
 after_success:
   - R CMD INSTALL $PKG_TARBALL
   - Rscript -e 'covr::codecov()'
