### 2018-10-23 70ee1b4
Added legend option to helical wheels
