### 2018-10-23 43c2575
Added legend option to wenxiang diagrams
