+++ R/draw_wenxiang.R
@@ -14,7 +14,8 @@
 #' draw_wenxiang("GIGAVLKVLTTGLPALIS")
 #' draw_wenxiang("QQRKRKIWSILAPLGTTL")
 draw_wenxiang <- function(sequence, col = c("grey", "yellow", "blue", "red"),
-                          labels = FALSE, label.col = "black", fixed = TRUE) {
+                          labels = FALSE, label.col = "black", fixed = TRUE,
+                          legend = FALSE) {
   # check length of sequence
   if (nchar(sequence) > 18) {
     stop("ERROR: sequence must have less than or equal to 18 characters.")
@@ -77,7 +78,8 @@ draw_wenxiang <- function(sequence, col = c("grey", "yellow", "blue", "red"),
                      FUN = function(curr.resid) {
                        residue_col(curr.resid)
                      })
-  df.resid$fill.col <- col[fill.col]
+  resid.types <- c("hydrophobic", "polar", "basic", "acidic")
+  df.resid$fill.col <- resid.types[fill.col]
   df.resid$lettername <- vapply(X = 1:nchar(sequence),
                                 FUN.VALUE = character(1),
                                 FUN = function(i) substr(sequence, i, i))
@@ -91,8 +93,13 @@ draw_wenxiang <- function(sequence, col = c("grey", "yellow", "blue", "red"),
                                         end = .data$end.angle)) +
          ggforce::geom_circle(data = df.resid,
                               ggplot2::aes(x0 = .data$x, y0 = .data$y,
-                                           r = 0.04, fill = I(.data$fill.col))) +
+                                           r = 0.04, fill = .data$fill.col)) +
          ggplot2::xlim(c(0, 1)) + ggplot2::ylim(c(0, 1)) +
+         ggplot2::scale_fill_manual(values = c("hydrophobic" = col[1],
+                                               "polar" = col[2],
+                                               "basic" = col[3],
+                                               "acidic" = col[4]),
+                                    name = "Residue Types") +
          ggplot2::theme(panel.grid.major = ggplot2::element_blank(),
                         panel.grid.minor = ggplot2::element_blank(),
                         panel.background = ggplot2::element_blank(),
@@ -109,8 +116,14 @@ draw_wenxiang <- function(sequence, col = c("grey", "yellow", "blue", "red"),
                                              label = .data$lettername,
                                              colour = I(label.col)))
   }
+  # fixed coordinates if user desires
   if (fixed) {
     g <- g + ggplot2::coord_fixed()
   }
+  # legend if user desires
+  if (legend) {
+    g <- g + ggplot2::theme(legend.position = "right")
+  }
+  
   g
 }
