### 2018-10-23 276b0e2
Added legend parameter to documentation (draw_wheel and draw_wenxiang)
