### 2018-10-23 rrrlw/helixvis#10 @rrrlw
wenxiang diagrams do not need to be limited to 18 residues
Following is a suggestion by @nagoue from [this JOSS review](https://github.com/openjournals/joss-reviews/issues/1008):

There is no reason to limit peptide length to 18 aa in wenxiang diagram unless to make it comparable to the helical diagram. I would rather suggest not limiting this length to wenxiang diagram style and "only" give a warning to the user to inform him that the comparison between the 2 diagram styles will not be possible.
