### 2018-10-23 #1008 @rrrlw
Thank you to the reviewers for their clear feedback.

In response to the points brought up by @nagoue:
1. The units (>= 0.6) dependency has been added to the Imports field in the DESCRIPTION file.
2. The typo in line 40 of paper.md has been fixed.
3. The redirection to Sandve et al. is not working for me either. However, the DOI in paper.md matches the DOI on the PLoS CB website (https://journals.plos.org/ploscompbiol/article?id=10.1371/journal.pcbi.1003285), which also does not redirect properly. Since the DOIs match, I have left it in paper.md unchanged; however, since redirection is not working, I would be willing to remove the DOI from the Sandve et al. reference in the JOSS paper.
4. Legends can be added to helical wheels (with `draw_wheel`) and wenxiang diagrams (with `draw_wenxiang`) using the `legend=TRUE` parameter. A corresponding section will be added in the helixvis introduction vignette.
5. You make a good point about limiting the wenxiang diagram to 18 residues. Because the code is currently based on explicit coordinates (that are based on a max of 18 residues), adding this feature would require major code changes. However, this is would be a very useful enhancement for users as wenxiang diagrams are not inherently limited in sequence length as helical wheels are, so I have added this as an [Issue](https://github.com/rrrlw/helixvis/issues/10) to the helixvis GitHub repo, and will work on it in the future to improve helixvis.

In response to the points brought up by @tomsing1:
1. [Issue 9](https://github.com/rrrlw/helixvis/issues/9) has been fixed by adding `build_vignettes = TRUE` to the GitHub README.md file.
2. The version mismatch has been corrected with the [latest helixvis release](https://github.com/rrrlw/helixvis/releases/tag/v1.0.0).
3. Thank you for the trust on authorship.
4. [Issue 7](https://github.com/rrrlw/helixvis/issues/7) has been fixed by removing the redundant files and only retaining the R script in the data directory.
5. lintr has been added to Travis CI for checking code style; notifications are all about snake case for variables (my personal style is to separate words in variables with a period and words in functions with an underscore) and not having greater than 80 characters of code in a single line (only error and warning messages seem to violate this). I do not think either of these types of errors are major code style violations.

Please let me know if there are any other changes I should make.
