### 2018-10-23 #1008 @tomsing1
Looks good to me - accept!
