### 2018-10-24 rrrlw/helixvis#8 @rrrlw
Fixed with [this release](https://github.com/rrrlw/helixvis/releases/tag/v1.0.0). Thank you!
