### 2018-10-24 80e7b70
Added testthat code coverage for plot legend enhancement
