### 2018-10-24 #1008 @yochannah
@nagoue and @tomsing1 - thanks for your great work reviewing! 👍 There's still once checkbox each that needs to be checked. Are we happy to check these and move ahead?
