### 2018-10-24 #1008 @tomsing1
Thanks for reminder - we are good to go!
