### 2018-10-25 0fd5cb8
Updated version number for Zenodo release
