+++ JOSS-paper/paper.bib
@@ -1,3 +1,11 @@
+@article{zenodo,
+  author = "R R Wadhwa and V Subramanian and R Stevens-Truss",
+  title = "rrrlw/helixvis: v1.0.1",
+  year = "2018",
+  doi = "10.5281/zenodo.1470939",
+  publisher = "Zenodo"
+}
+
 @article{wheeldiag,
   author = "M Schiffer and A B Edmundson",
   title = "Use of helical wheels to represent the structures of proteins and to identify segments with helical potential",
