### 2018-10-25 67494f3
Added zenodo citation to JOSS paper
