### 2018-10-25 2406aac
Added Zenodo badge to README
