### 2018-10-25 05b4271
Added indication for who senior author of JOSS paper is
