### 2018-10-25 f997a68
Added clarification to JOSS paper
