### 2018-10-25 #1008 @rrrlw
Thank you, @tomsing1! I have made a couple of changes to the JOSS paper text so will recompile here to get an updated PDF, my apologies for any resulting notification spam.
