### 2018-10-28 #1008 @rrrlw
Hi @nagoue, thank you for reviewing helixvis! Is there anything else you would like me to modify to satisfy the last checkbox?
