### 2018-11-04 32ec9ab
Added JOSS DOI badge to README
