### 2018-11-04 2d9813f
Removed units dependency to fix R CMD check NOTE
