+++ DESCRIPTION
@@ -16,8 +16,7 @@ LazyData: true
 Depends: R (>= 2.10)
 Imports: ggplot2 (>= 3.0.0),
   ggforce (>= 0.1.3),
-  rlang (>= 0.2.2),
-  units (>= 0.6)
+  rlang (>= 0.2.2)
 URL: https://github.com/rrrlw/helixvis
 BugReports: https://github.com/rrrlw/helixvis/issues
 Suggests: knitr,
