### 2018-11-04 3179f7e
Updated cran-comments for upcoming CRAN submission of v1.0.1
