+++ cran-comments.md
@@ -1,10 +1,12 @@
 ## Test environments
-* local Ubuntu 18.04 install, R 3.4.4
+* local Ubuntu 18.04 install, r-release
 * Travis CI (Ubuntu 14.04), r-release
-* r-hub Windows Server 2008 R2, r-devel
+* AppVeyor CI (Windows Server 2012 R2 x64), r-release
+* r-hub (Linux), r-devel
+* r-hub (Mac OS), r-release
 
 ## R CMD check results
 There were no ERRORs, WARNINGs, or NOTEs.
 
 ## Special notes
-Version 0.9.1 allows users to visualize helices as Wenxiang diagrams.
+helixvis v1.0.1 has been accepted for publication in the Journal of Open Source Software.
