### 2018-11-04 #1008 @rrrlw
Thank you so much @nagoue and @tomsing1 for reviewing, @yochannah for editing, and @arfon for setting up a great experience with JOSS!
