### 2018-11-04 #1008 @arfon
@rrrlw - your paper is now accepted into JOSS ⚡️🚀💥
